requirements.txt
pymongo==4.7.2
qdrant-client==1.9.1
sentence-transformers==2.7.0
schedule==1.2.2
fastapi==0.111.0
uvicorn==0.30.1
openai==0.28.0
python-dotenv==1.0.1